+++ packages/in_app_purchase/in_app_purchase_ios/example/ios/RunnerTests/Stubs.h
@@ -60,4 +60,9 @@ API_AVAILABLE(ios(11.2), macos(10.13.2))
 - (instancetype)initWithFailureError:(NSError *)error;
 @end
 
+API_AVAILABLE(ios(13.0), macos(10.15))
+@interface SKStorefrontStub : SKStorefront
+- (instancetype)initWithMap:(NSDictionary *)map;
+@end
+
 NS_ASSUME_NONNULL_END